chore(#293): move nuxt module files
