+++ packages/views/src/index.js
@@ -1,4 +1,4 @@
-import { DruxtViewsNuxtModule } from './nuxtModule'
+import { DruxtViewsNuxtModule } from './nuxt'
 
 /**
  * The Nuxt.js module function.
@@ -8,7 +8,7 @@ import { DruxtViewsNuxtModule } from './nuxtModule'
  * @type {Function}
  * @exports default
  * @name DruxtViewsNuxtModule
- * @see {@link ./module|DruxtViewsNuxtModule}
+ * @see {@link ./nuxt|DruxtViewsNuxtModule}
  */
 export default DruxtViewsNuxtModule
